# swagger/doc.yml
swagger: "2.0"
info:
  description: "A GinBro RESTful APIs"
  version: "1.0.0"
  title: "GinBro RESTful APIs Application"
host: "127.0.0.1:5555"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /g-role:
    get:
      tags:
      - "g-role"
      summary: "get all g-role by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GRolePagination"
    post:
      tags:
      - "g-role"
      summary: "create g-role"
      description: "create g-role"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-role"
        required: true
        schema:
          $ref: "#/definitions/GRole"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "g-role"
      summary: "update g-role"
      description: "update g-role"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-role"
        required: true
        schema:
          $ref: "#/definitions/GRole"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-role/{Id}:
    get:
      tags:
      - "g-role"
      summary: "get a g-role by ID"
      description: "get a g-role by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-role to return"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GRole"
    delete:
      tags:
      - "g-role"
      summary: "Destroy a g-role by ID"
      description: "delete a g-role by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-role to delete"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-role-menu:
    get:
      tags:
      - "g-role-menu"
      summary: "get all g-role-menu by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GRoleMenuPagination"
    post:
      tags:
      - "g-role-menu"
      summary: "create g-role-menu"
      description: "create g-role-menu"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-role-menu"
        required: true
        schema:
          $ref: "#/definitions/GRoleMenu"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "g-role-menu"
      summary: "update g-role-menu"
      description: "update g-role-menu"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-role-menu"
        required: true
        schema:
          $ref: "#/definitions/GRoleMenu"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-role-menu/{Id}:
    get:
      tags:
      - "g-role-menu"
      summary: "get a g-role-menu by ID"
      description: "get a g-role-menu by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-role-menu to return"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GRoleMenu"
    delete:
      tags:
      - "g-role-menu"
      summary: "Destroy a g-role-menu by ID"
      description: "delete a g-role-menu by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-role-menu to delete"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-user:
    get:
      tags:
      - "g-user"
      summary: "get all g-user by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GUserPagination"
    post:
      tags:
      - "g-user"
      summary: "create g-user"
      description: "create g-user"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-user"
        required: true
        schema:
          $ref: "#/definitions/GUser"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "g-user"
      summary: "update g-user"
      description: "update g-user"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-user"
        required: true
        schema:
          $ref: "#/definitions/GUser"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-user/{Id}:
    get:
      tags:
      - "g-user"
      summary: "get a g-user by ID"
      description: "get a g-user by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-user to return"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GUser"
    delete:
      tags:
      - "g-user"
      summary: "Destroy a g-user by ID"
      description: "delete a g-user by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-user to delete"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
  /abnormalattendance:
    get:
      tags:
      - "abnormalattendance"
      summary: "get all abnormalattendance by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AbnormalattendancePagination"
    post:
      tags:
      - "abnormalattendance"
      summary: "create abnormalattendance"
      description: "create abnormalattendance"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create abnormalattendance"
        required: true
        schema:
          $ref: "#/definitions/Abnormalattendance"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "abnormalattendance"
      summary: "update abnormalattendance"
      description: "update abnormalattendance"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create abnormalattendance"
        required: true
        schema:
          $ref: "#/definitions/Abnormalattendance"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /abnormalattendance/{Id}:
    get:
      tags:
      - "abnormalattendance"
      summary: "get a abnormalattendance by ID"
      description: "get a abnormalattendance by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of abnormalattendance to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Abnormalattendance"
    delete:
      tags:
      - "abnormalattendance"
      summary: "Destroy a abnormalattendance by ID"
      description: "delete a abnormalattendance by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of abnormalattendance to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
definitions:
...
  GRole:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: ""
        format: "int32"
      
      created_at:
        type: "string"
        description: ""
        format: "date-time"
      
      updated_at:
        type: "string"
        description: ""
        format: "date-time"
      
      deleted_at:
        type: "string"
        description: ""
        format: "date-time"
      
      record_id:
        type: "string"
        description: ""
        format: "string"
      
      name:
        type: "string"
        description: ""
        format: "string"
      
      sequence:
        type: "integer"
        description: ""
        format: "int64"
      
      memo:
        type: "string"
        description: ""
        format: "string"
      
      creator:
        type: "string"
        description: ""
        format: "string"
  GRolePagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/GRole"
  GRoleMenu:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: ""
        format: "int32"
      
      created_at:
        type: "string"
        description: ""
        format: "date-time"
      
      updated_at:
        type: "string"
        description: ""
        format: "date-time"
      
      deleted_at:
        type: "string"
        description: ""
        format: "date-time"
      
      role_id:
        type: "string"
        description: ""
        format: "string"
      
      menu_id:
        type: "string"
        description: ""
        format: "string"
      
      action:
        type: "string"
        description: ""
        format: "string"
      
      resource:
        type: "string"
        description: ""
        format: "string"
  GRoleMenuPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/GRoleMenu"
  GUser:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: ""
        format: "int32"
      
      created_at:
        type: "string"
        description: ""
        format: "date-time"
      
      updated_at:
        type: "string"
        description: ""
        format: "date-time"
      
      deleted_at:
        type: "string"
        description: ""
        format: "date-time"
      
      record_id:
        type: "string"
        description: ""
        format: "string"
      
      user_name:
        type: "string"
        description: ""
        format: "string"
      
      real_name:
        type: "string"
        description: ""
        format: "string"
      
      password:
        type: "string"
        description: ""
        format: "string"
      
      email:
        type: "string"
        description: ""
        format: "string"
      
      phone:
        type: "string"
        description: ""
        format: "string"
      
      status:
        type: "integer"
        description: ""
        format: "int64"
      
      creator:
        type: "string"
        description: ""
        format: "string"
  GUserPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/GUser"
...
  Abnormalattendance:
    type: "object"
    properties:
    
      id:
        type: "string"
        description: "ID"
        format: "string"
      
      tenant_id:
        type: "string"
        description: "租户编号"
        format: "string"
      
      person_id:
        type: "string"
        description: "雇员编号"
        format: "string"
      
      att_type:
        type: "string"
        description: "考勤类别"
        format: "string"
      
      att_name:
        type: "string"
        description: "类别名称"
        format: "string"
      
      start_time:
        type: "string"
        description: "开始时间"
        format: "date-time"
      
      end_time:
        type: "string"
        description: "结束时间"
        format: "date-time"
      
      amount:
        type: "number"
        description: "数值"
        format: "float"
      
      biz_no:
        type: "string"
        description: "单号"
        format: "string"
      
      fee:
        type: "number"
        description: "考勤扣费"
        format: "float"
      
      rule_name:
        type: "string"
        description: "扣费规则"
        format: "string"
      
      formula_text:
        type: "string"
        description: "扣费公式"
        format: "string"
      
      f_enabled_mark:
        type: "integer"
        description: "有效标志0否1是"
        format: "int64"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  AbnormalattendancePagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Abnormalattendance"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
...